Split type declarations
